Star Map Implementation
Clicking 'M' now brings up the star map, which consists of small versions of however many stars that are intended to load. Eventually these small "generic stars" will be able to be clicked to bring the player to that system. Additionally, changed how systems are loaded so that all systems are loaded beforehand, shortening times between initial system transfers and enabling information about stars to be transmitted to the star map.
However, the current system of switching between star systems is faulty as the player is able to interact with the stars/planets of all loaded systems at any point, including in the star map. This needs to be fixed