Added Home System
The system with id 0 is now designated as the 'Home System,' or the starting system of the game. This system has either a K, G, or F class star, a water world in the habitable zone of the star, at least one gas giant, 2 asteroid belts, and between 7 and 9 planets. Also, this system now appears at the center of the star map.